Fix errore tipo variabile

+++ backend/src/main/java/com/immobiliaris/demo/entity/User.java
@@ -15,7 +15,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "Id_utente")
-    private Integer idUtente;
+    private Long idUtente;
 
     @Column(name = "CF", length = 16)
     private String cf;
